requirements.txt
gunicorn 
Flask==2.4.1
python-dotenv==1.0.0
PyJWT==2.8.0